requirements.txt
asgiref==3.6.0
boto3==1.24.69
botocore==1.27.69
certifi==2022.6.15
charset-normalizer==2.1.0
# cloudinary==1.29.0
# dj3-cloudinary-storage==0.0.6
Django==4.2.1
# django-crispy-forms==2.0
# django-embed-video==1.4.8
django-storages==1.12.3
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
Pillow==9.2.0
psycopg2==2.9.3
# psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.10
# whitenoise==6.2.0
